Refactor Databricks Delta transaction support
