+++ delta-lake/common/src/main/db-350db143-400db173/scala/com/nvidia/spark/rapids/delta/GpuDeltaDataSource.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2026, NVIDIA CORPORATION.
+ * Copyright (c) 2024-2026, NVIDIA CORPORATION.
  *
  * This file was derived from DeltaDataSource.scala in the
  * Delta Lake project at https://github.com/delta-io/delta.
