+++ delta-lake/common/src/main/db-350db143-400db173/scala/com/nvidia/spark/rapids/delta/shims/DeltaLogShim.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, NVIDIA CORPORATION.
+ * Copyright (c) 2024-2026, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
