+++ delta-lake/common/src/main/db-350db143-400db173/scala/com/nvidia/spark/rapids/delta/shims/MetadataShims.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2026, NVIDIA CORPORATION.
+ * Copyright (c) 2024-2026, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,20 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
 package com.nvidia.spark.rapids.delta.shims
 
-import com.databricks.sql.transaction.tahoe.DeltaLog
-import com.databricks.sql.transaction.tahoe.actions.Metadata
+import com.databricks.sql.transaction.tahoe.stats.DeltaStatistics
 
-import org.apache.spark.sql.execution.datasources.FileFormat
-
-object DeltaLogShim {
-  def fileFormat(deltaLog: DeltaLog): FileFormat = {
-    deltaLog.fileFormat(deltaLog.unsafeVolatileSnapshot.protocol,
-      deltaLog.unsafeVolatileSnapshot.metadata)
-  }
-  def getMetadata(deltaLog: DeltaLog): Metadata = {
-    deltaLog.unsafeVolatileSnapshot.metadata
-  }
+trait ShimUsesMetadataFields {
+  val NUM_RECORDS = DeltaStatistics.NUM_RECORDS
+  val MIN = DeltaStatistics.MIN
+  val MAX = DeltaStatistics.MAX
+  val NULL_COUNT = DeltaStatistics.NULL_COUNT
 }
