+++ delta-lake/common/src/main/db-350db143-400db173/scala/com/nvidia/spark/rapids/delta/shims/ShimDeltaUDF.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, NVIDIA CORPORATION.
+ * Copyright (c) 2024-2026, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.nvidia.spark.rapids.delta.shims
 
-import com.databricks.sql.transaction.tahoe.constraints.Constraints._
-import com.databricks.sql.transaction.tahoe.schema.DeltaInvariantViolationException
-import com.databricks.sql.transaction.tahoe.schema.InvariantViolationException
+import com.databricks.sql.transaction.tahoe.DeltaUDF
 
-object InvariantViolationExceptionShim {
-  def apply(c: Check, m: Map[String, Any]): InvariantViolationException = {
-    DeltaInvariantViolationException(c, m)
-  }
+import org.apache.spark.sql.expressions.UserDefinedFunction
 
-  def apply(c: NotNull): InvariantViolationException = {
-    DeltaInvariantViolationException(c)
-  }
+object ShimDeltaUDF {
+  def stringStringUdf(f: String => String): UserDefinedFunction = DeltaUDF.stringFromString(f)
 }
