+++ delta-lake/delta-spark350db143/src/main/scala/com/databricks/sql/transaction/tahoe/rapids/GpuWriteIntoDelta.scala
@@ -22,52 +22,19 @@
 package com.databricks.sql.transaction.tahoe.rapids
 
 import com.databricks.sql.transaction.tahoe.{DeltaOperations, OptimisticTransaction}
+import com.databricks.sql.transaction.tahoe.actions.Action
 import com.databricks.sql.transaction.tahoe.commands.WriteIntoDeltaEdge
-
-import org.apache.spark.sql._
-import org.apache.spark.sql.execution.command.LeafRunnableCommand
 
 /** GPU version of Delta Lake's WriteIntoDelta. */
 case class GpuWriteIntoDelta(
     gpuDeltaLog: GpuDeltaLog,
     cpuWrite: WriteIntoDeltaEdge)
-    extends LeafRunnableCommand {
+    extends GpuWriteIntoDeltaBase(gpuDeltaLog, cpuWrite) {
 
-  override def run(sparkSession: SparkSession): Seq[Row] = {
-    gpuDeltaLog.withNewTransaction { txn =>
-      // If this batch has already been executed within this query, then return.
-      val skipExecution = hasBeenExecuted(txn)
-      if (skipExecution) {
-        return Seq.empty
-      }
-
-      val actions = cpuWrite.write(txn, sparkSession)
-      val operation = DeltaOperations.Write(
-        cpuWrite.mode,
-        Option(cpuWrite.partitionColumns),
-        cpuWrite.options.replaceWhere,
-        cpuWrite.options.userMetadata)
-      txn.commit(actions, operation)
-    }
-    Seq.empty
-  }
-
-  /**
-   * Returns true if there is information in the spark session that indicates that this write, which
-   * is part of a streaming query and a batch, has already been successfully written.
-   */
-  private def hasBeenExecuted(txn: OptimisticTransaction): Boolean = {
-    val txnVersion = cpuWrite.options.txnVersion
-    val txnAppId = cpuWrite.options.txnAppId
-    for (v <- txnVersion; a <- txnAppId) {
-      val currentVersion = txn.txnVersion(a)
-      if (currentVersion >= v) {
-        logInfo(s"Transaction write of version $v for application id $a " +
-            s"has already been committed in Delta table id ${txn.deltaLog.tableId}. " +
-            s"Skipping this write.")
-        return true
-      }
-    }
-    false
+  override protected def commitActions(
+      txn: OptimisticTransaction,
+      actions: Seq[Action],
+      operation: DeltaOperations.Operation): Unit = {
+    txn.commit(actions, operation)
   }
 }
