+++ delta-lake/delta-spark400db173/pom.xml
@@ -68,6 +68,7 @@
                             <sources>
                                 <source>${spark.rapids.source.basedir}/delta-lake/common/src/main/scala</source>
                                 <source>${spark.rapids.source.basedir}/delta-lake/common/src/main/databricks/scala</source>
+                                <source>${spark.rapids.source.basedir}/delta-lake/common/src/main/db-350db143-400db173/scala</source>
                             </sources>
                         </configuration>
                     </execution>
