+++ scala2.13/delta-lake/delta-spark350db143/pom.xml
@@ -65,6 +65,7 @@
                             <sources>
                                 <source>${project.basedir}/../common/src/main/scala</source>
                                 <source>${project.basedir}/../common/src/main/databricks/scala</source>
+                                <source>${project.basedir}/../common/src/main/db-350db143-400db173/scala</source>
                             </sources>
                         </configuration>
                     </execution>
